Move settings and stats into /player
