+++ components/app/video/player/video-stats-panel.tsx
